pyproject.toml
[project]
name = "fastmcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio>=4.10.0",
    "fastmcp>=2.11.3",
    "httpx>=0.28.1",
    "pyodide-py>=0.27.7",
]
